# omnivoice-dify-gateway/openapi/dify-openapi.yaml
openapi: 3.0.3
info:
  title: Shenxiang OmniVoice Dify Gateway
  version: 0.1.0
  description: JSON-first speech generation gateway for Dify and shenxiang.school.
servers:
  - url: https://voice-api.shenxiang.school
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Voice gateway API key.
  schemas:
    VoiceInfo:
      type: object
      properties:
        voice_id:
          type: string
          description: Voice identifier accepted by the gateway.
        name:
          type: string
          description: Human friendly voice name.
        language:
          type: string
          description: Primary language for the voice.
        description:
          type: string
          description: Recommended usage for this voice.
        enabled:
          type: boolean
          description: Whether the voice can currently be used.
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine readable error code.
            message:
              type: string
              description: Safe human readable error message.
            request_id:
              type: string
              description: Request ID for log correlation.
paths:
  /healthz:
    get:
      operationId: healthCheck
      security: []
      summary: Check gateway and OmniVoice health.
      responses:
        "200":
          description: Health status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, description: Overall status. }
                  gateway: { type: string, description: Gateway status. }
                  omnivoice: { type: string, description: Upstream status, ok or unreachable. }
                  version: { type: string, description: Gateway version. }
                  time: { type: string, description: ISO timestamp. }
  /v1/voices:
    get:
      operationId: listVoices
      summary: List enabled voices.
      responses:
        "200":
          description: Voice list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  voices:
                    type: array
                    description: Voices available to Dify.
                    items:
                      $ref: "#/components/schemas/VoiceInfo"
  /v1/tts:
    post:
      operationId: createTTS
      summary: Generate speech from text.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: Text to synthesize. Limited by MAX_TEXT_CHARS.
                voice_id:
                  type: string
                  description: Allowed voice ID, for example teacher_female_01.
                  default: teacher_female_01
                language:
                  type: string
                  description: Language code.
                  default: zh-CN
                speed:
                  type: number
                  description: Speech speed multiplier.
                  default: 1.0
                emotion:
                  type: string
                  description: Style or emotion hint.
                  default: friendly
                format:
                  type: string
                  description: Output format.
                  enum: [mp3, wav, flac, opus]
                  default: mp3
                return_mode:
                  type: string
                  description: JSON URL return mode for Dify.
                  enum: [url]
                  default: url
                sync:
                  type: boolean
                  description: Whether to wait for audio generation.
                  default: false
      responses:
        "200":
          description: Queued or completed TTS result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string, description: Gateway job ID. }
                  status: { type: string, description: queued, running, succeeded, or failed. }
                  audio_url: { type: string, description: Public audio URL when completed. }
                  filename: { type: string, description: Media filename. }
                  duration_seconds: { type: number, description: Audio duration when available. }
        "400":
          description: Invalid request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  /v1/essay-commentary:
    post:
      operationId: createEssayCommentary
      summary: Generate AI teacher essay commentary audio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grading_report]
              properties:
                student_grade:
                  type: string
                  description: Student grade, e.g. 初中.
                essay_title:
                  type: string
                  description: Essay title.
                essay_text:
                  type: string
                  description: Full student essay. The gateway will not read long essays verbatim.
                grading_report:
                  type: string
                  description: Existing grading report or commentary source text.
                voice_id:
                  type: string
                  description: Allowed voice ID.
                  default: teacher_female_01
                style:
                  type: string
                  description: Commentary speaking style.
                duration_target_seconds:
                  type: integer
                  description: Desired target duration.
                sync:
                  type: boolean
                  description: Whether to wait for audio generation.
                  default: false
      responses:
        "200":
          description: Commentary text and queued/completed audio result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string, description: Gateway job ID. }
                  status: { type: string, description: queued or succeeded. }
                  commentary_text: { type: string, description: Text sent to TTS. }
                  audio_url: { type: string, description: Public audio URL when completed. }
  /v1/jobs/{job_id}:
    get:
      operationId: getJobStatus
      summary: Get queued speech generation job status.
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: string }
          description: Job ID returned by createTTS or createEssayCommentary.
      responses:
        "200":
          description: Job status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string, description: Gateway job ID. }
                  status: { type: string, description: queued, running, succeeded, or failed. }
                  progress: { type: number, description: Progress between 0 and 1. }
                  audio_url: { type: string, description: Public audio URL when completed. }
                  error: { type: string, description: Safe error message if failed. }
                  created_at: { type: string, description: Creation timestamp. }
                  updated_at: { type: string, description: Last update timestamp. }
  /v1/voice-clone:
    post:
      operationId: cloneVoice
      summary: Create a restricted voice clone profile.
      description: Disabled by default. Requires consent_confirmed=true and administrator review.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [speaker_name, consent_confirmed, intended_use]
              properties:
                audio_url:
                  type: string
                  description: Public HTTP or HTTPS reference audio URL. Private/internal IPs are blocked.
                base64_audio:
                  type: string
                  description: Base64 reference audio. Use only when audio_url is unavailable.
                speaker_name:
                  type: string
                  description: Speaker name for audit.
                consent_confirmed:
                  type: boolean
                  description: Must be true. Confirms speaker consent and rights.
                intended_use:
                  type: string
                  description: Declared intended use.
                owner_contact:
                  type: string
                  description: Optional contact for audit.
      responses:
        "200":
          description: Clone result or pending review.
          content:
            application/json:
              schema:
                type: object
                properties:
                  voice_id: { type: string, description: Created voice ID. }
                  status: { type: string, description: disabled, created, or pending_admin_review. }
                  message: { type: string, description: Human readable result. }
        "403":
          description: Clone disabled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorEnvelope" }